backend-python/pyproject.toml
[tool.poetry]
name = "brex-interview-playground-backend"
version = "0.1.0"
description = "Brex Interview Playground Backend (Python)"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
strawberry-graphql = {extras = ["fastapi"], version = "^0.278.1"}
pydantic = {extras = ["email"], version = "^2.6.3"}
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3 